+++ examples/run_notebook.sh
@@ -1,0 +1,1 @@
+jupyter nbconvert --execute --to notebook --inplace exploration.ipynb --ExecutePreprocessor.timeout=-1