+++ examples/run_py.sh
@@ -1,0 +1,4 @@
+# python -u exploration.py 2>&1 | tee output.txt
+python -u exploration.py 2 ;\
+python -u exploration.py 5 ;\
+python -u exploration.py 10