+++ inept/utilities.py
@@ -1,3 +1,7 @@
+from time import perf_counter
+import tracemalloc
+
+import numpy as np
 import torch
 
 
@@ -12,3 +16,119 @@ def cosine_similarity(a):
 def euclidean_distance(a):
     # Calculate euclidean distance
     return torch.cdist(a, a, p=2)
+
+
+class time_logger():
+    """Class made for easy logging with toggleable verbosity"""
+    def __init__(
+        self,
+        discard_first_sample=False,
+        record=True,
+        verbose=False,
+        memory_usage=False,
+    ):
+        self.discard_first_sample = discard_first_sample
+        self.record = record
+        self.verbose = verbose
+        self.memory_usage = memory_usage
+
+        self.history = {}
+        self.start_time = perf_counter()
+
+        if memory_usage:
+            self.history_mem = {}
+            tracemalloc.start()
+
+    def log(self, str=''):
+        """Print with message ``str`` if verbose.  Otherwise, skip"""
+        if not (self.verbose or self.record):
+            return  # Cut timing for optimization
+        self.end_time = perf_counter()
+
+        # Perform any auxiliary operations here
+        time_elapsed = self.end_time - self.start_time
+        # Record time
+        if self.record:
+            if str not in self.history:
+                self.history[str] = []
+            self.history[str].append(time_elapsed)
+        if self.verbose:
+            print(f'{str}: {time_elapsed}')
+        # Record memory
+        if self.memory_usage:
+            if self.record:
+                if str not in self.history_mem:
+                    self.history_mem[str] = []
+                self.history_mem[str].append(tracemalloc.get_traced_memory())
+            if self.verbose:
+                print(f'{str} Memory: Stored {self.history_mem[-1][0]} - Peak {self.history_mem[-1][1]}')
+            tracemalloc.stop()
+
+        # Re-time to avoid extra runtime cost
+        self.start_time = perf_counter()
+        if self.memory_usage:
+            # WARNING, does not end tracemalloc
+            tracemalloc.start()
+
+    def aggregate(self, method='mean'):
+        """Print mean times for all keys in ``self.history``"""
+        running_total = 0
+        for k, v in self.history.items():
+            avg_time_elapsed = np.array(v)
+            if self.discard_first_sample:
+                avg_time_elapsed = avg_time_elapsed[1:]
+            if method == 'mean':
+                avg_time_elapsed = np.mean(np.array(v))
+            elif method == 'sum':
+                avg_time_elapsed = np.sum(np.array(v))
+
+            running_total += avg_time_elapsed
+            print(f'{k}: {avg_time_elapsed}')
+            if self.memory_usage:
+                stored = 0
+                peak = 0
+                for val in self.history_mem[k]:
+                    stored += val[0]
+                    if val[1] > peak:
+                        peak = val[1]
+                print(f'{k} Memory: Stored {stored} - Peak {peak}')
+        print(f'Total: {running_total}')
+
+
+class EarlyStopping:
+    def __init__(self, buffer=50, delta=.01, decreasing=False):
+        self.buffer = buffer
+        self.delta = delta
+        self.decreasing = decreasing
+
+        # Initialize
+        self.reset()
+
+    def __call__(self, objective):
+        """Return `True` if stop, else `False`."""
+        # First entry
+        if self.best is None: self.set_best(objective); return False
+
+        # Check for new best
+        if self.decreasing and (objective < self.threshold): self.set_best(objective)
+        elif not self.decreasing and (objective > self.threshold): self.set_best(objective)
+        else: self.lapses += 1
+
+        # Check lapses
+        if self.lapses >= self.buffer:
+            return True
+        return False
+
+    def reset(self):
+        """Reset class to default state"""
+        self.best = None
+        self.threshold = None
+        self.lapses = 0
+
+    def set_best(self, objective):
+        self.best = objective
+        self.calculate_threshold()
+        self.lapses = 0
+
+    def calculate_threshold(self):
+        self.threshold = self.best + (-1 if self.decreasing else 1) * self.delta